Subject: update of import and data check
Added and checked the various import that we could use for this project
in the 1st cell and added a number of data (image) check to understand
the size of the dataset we have. See as below

There are 2750 total images.
There are 2000 total train images.
There are 150 total validation images.
There are 600 total test images.