subject: Completion 1st step of Dermato_AI
Completion of the model using transfer learning with VGG16. I was not
able to train the model well enough and use GPU to have a decent
accuracy, but all the elements are here. The ROC diagram is also
available and I tested the model on specific Melanoma, but it doesn't
recognise it. So the model will need to be retrained.