Style & test: update the tests and train model on 2 epochs
I cleaned several elements from the previous commit to make it easier
to read and I trained the model on 2 epochs. The training loss is
improved but still not good enough as we tested the model on testset,
which was composed of melanoma pictures only, and it was not able to
recognize any yet.